[ADD] if guard in case folder does not exist when setting the scripts

+++ setup.sh
@@ -16,22 +16,30 @@ test -e "/etc/my_common/includes" || {
     mkdir /etc/my_common/includes && chmod 755 /etc/my_common/includes
 }
 
-for ele in $(ls ./scripts); do
-    cp ./scripts/$ele /etc/my_common/bin/$ele
-    chmod 755 /etc/my_common/bin/$ele
-done
+if [ -e "./scripts" ]; then
+    for ele in $(ls ./scripts); do
+        cp ./scripts/$ele /etc/my_common/bin/$ele
+        chmod 755 /etc/my_common/bin/$ele
+    done
+fi
 
-for ele in $(ls ./sources); do
-    cp ./sources/$ele /etc/my_common/sources/$ele
-    chmod 755 /etc/my_common/sources/$ele
-done
+if [ -e "./sources" ]; then
+    for ele in $(ls ./sources); do
+        cp ./sources/$ele /etc/my_common/sources/$ele
+        chmod 755 /etc/my_common/sources/$ele
+    done
+fi
 
-for ele in $(ls ./links); do
-    cp ./links/$ele /etc/profile.d/$ele
-    chmod 644 /etc/profile.d/$ele
-done
+if [ -e "./links" ]; then
+    for ele in $(ls ./links); do
+        cp ./links/$ele /etc/profile.d/$ele
+        chmod 644 /etc/profile.d/$ele
+    done
+fi
 
-for ele in $(ls ./includes); do
-    cp ./includes/$ele /etc/my_common/includes/$ele
-    chmod 755 /etc/my_common/includes/$ele
-done
+if [ -e "./includes" ]; then
+    for ele in $(ls ./includes); do
+        cp ./includes/$ele /etc/my_common/includes/$ele
+        chmod 755 /etc/my_common/includes/$ele
+    done
+fi
